Hide word selector when an item is selected or a valid word is fully entered. Avoid error when no items in word selector and scrollToIndex 0.  Clean up log statements. Move usePrevious to its own file.
