+++ src/components/shared/usePrevious.ts
@@ -1,0 +1,9 @@
+import { useRef, useEffect } from 'react';
+
+export default function usePrevious(value: any) {
+  const ref = useRef();
+  useEffect(() => {
+    ref.current = value; //assign the value of ref to the argument
+  },[value]); //this code will run when the value of 'value' changes
+  return ref.current; //in the end, return the current ref value.
+}
